Multiple changes
* aPSKdLength type changed from uint32_t to uint8_t
* Added parameter data structure to IOCTL_OTLWF_OT_COMMISIONER_ADD_JOINER
* Reworked PSKd related definitions
* Removed string PSKd decoding to binary PSKd - all JoinerStart and AddJoiner user interfaces treat PSK string as a raw PSKd.

+++ include/openthread/commissioner.h
@@ -101,7 +101,7 @@ OTAPI otError OTCALL otCommissionerStop(otInstance *aInstance);
  *
  */
 OTAPI otError OTCALL otCommissionerAddJoiner(otInstance *aInstance, const otExtAddress *aExtAddress,
-                                             const uint8_t *aPSKd, uint32_t aPSKdLength, uint32_t aTimeout);
+                                             const uint8_t *aPSKd, uint8_t aPSKdLength, uint32_t aTimeout);
 
 /**
  * This function removes a Joiner entry.
