+++ src/core/api/joiner_api.cpp
@@ -39,7 +39,7 @@
 
 using namespace ot;
 
-otError otJoinerStart(otInstance *aInstance, const uint8_t *aPSKd, uint32_t aPSKdLength,
+otError otJoinerStart(otInstance *aInstance, const uint8_t *aPSKd, uint8_t aPSKdLength,
                       const char *aProvisioningUrl,
                       const char *aVendorName, const char *aVendorModel,
                       const char *aVendorSwVersion, const char *aVendorData,
